More testing. Use `sorted` argument of `np.unique`.

+++ src/hats/pixel_math/sparse_histogram.py
@@ -164,7 +164,7 @@ def supplemental_count_histogram(mapped_pixels, supplemental_count, highest_orde
     if supplemental_count is not None:
         if len(supplemental_count) != len(mapped_pixels):
             raise ValueError("mapped pixels and supplemental counts must be the same length")
-        supplemental_sums = np.zeros(len(mapped_pixel))
+        supplemental_sums = np.zeros(len(mapped_pixel), dtype=np.int64)
 
         for index, mem_size in zip(unique_inverse, supplemental_count, strict=True):
             supplemental_sums[index] += mem_size
